backend/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Utilidades
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Impresora térmica
python-escpos==3.0
Pillow>=10.0.0
pyusb>=1.2.1
# pywin32>=306  ← COMENTADO: Solo para Windows, no funciona en Docker/Linux

# Backups y exportación
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.1.3
openpyxl==3.1.2